+++ cpp/practice/StPetersburgParadox/StPetersburgParadox.h
@@ -2,4 +2,4 @@
 #include <cinttypes>
 
 int64_t doGambleAndGetResult(std::mt19937_64 &engine);
-int64_t getAverageResultOfGames(int numberOfTries);
+int64_t getAverageResultOfGames(int64_t numberOfTries);
